Add alloc uses
